requirements.txt
fastapi>=0.110.0,<0.120
uvicorn[standard]>=0.27.0,<0.30
httpx>=0.27.0,<0.28
pandas>=2.2.0,<2.3
scikit-learn>=1.4.0,<1.5
numpy>=1.26.0,<2.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

